A bit of PXX2 header cleanup
